backend/requirements.txt
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
fastapi[standard]==0.115.5
alembic==1.14.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pydantic_settings==2.6.1